Created folder for third page
Third folder created for third page that contains the page for the CSE Store User reports. Currently is only having template of the page and will need to add placeholder boxes as well as later on, functionality to the page. This is small commit. Next commit will have added placeholder boxes to it.